refactor(trade): 将订单 ID 类型从 int 改为 int64
- 在 GetOrderResponse 结构中，将 OrderId 字段的类型从 int 改为 int64
- 在 NewAllOrdersResponse 结构中，同样将 OrderId 字段的类型从 int 改为 int64
- 此修改是为了确保订单 ID 的完整性和防止溢出，提高系统稳定性

+++ trade.go
@@ -865,7 +865,7 @@ func (s *GetOrderService) Do(ctx context.Context, opts ...RequestOption) (res *G
 
 type GetOrderResponse struct {
 	ClientOrderId string       `json:"clientOrderId"`
-	OrderId       int          `json:"orderId"`
+	OrderId       int64        `json:"orderId"`
 	Symbol        string       `json:"symbol"`
 	Side          Side         `json:"side"`
 	PositionSide  PositionSide `json:"positionSide"`
@@ -965,7 +965,7 @@ func (s *GetAllOrdersService) Do(ctx context.Context, opts ...RequestOption) (re
 // Create NewAllOrdersResponse
 type NewAllOrdersResponse struct {
 	ClientOrderId string       `json:"clientOrderId"`
-	OrderId       int          `json:"orderId"`
+	OrderId       int64        `json:"orderId"`
 	Symbol        string       `json:"symbol"`
 	Side          Side         `json:"side"`
 	PositionSide  PositionSide `json:"positionSide"`
